fut_trader_fcflips_clone/requirements.txt
discord.py==2.3.2
requests
undetected-chromedriver
beautifulsoup4
python-dotenv